Implement Rule of Five and deep copy logic for Patient class
- Memory Management: Implemented manual heap allocation for the Treatment array in the Patient class to satisfy no-smart-pointer constraints.
- Deep Copy (TC-06): Added Copy Constructor and Copy Assignment to ensure independent patient records via manual cloning of treatments.
- Move Semantics (TC-07): Implemented Move Constructor and Move Assignment using pointer stealing and source nullification for efficient discharge.
- Base Class Integration: Integrated Person base class initialization and assignment within Patient methods.
- Data Integrity: Swapped to long long int for phone numbers to handle 11-digit formats and prevent overflow.

+++ Person.h
@@ -17,12 +17,12 @@ class Person {
     Date DOB;
     
     int ID;
-    int Phone;
+    long long int Phone;
     
     public:
     
     
-    Person(string name, Date dob, int ID, int Phone): name(name), DOB(dob), ID(ID), Phone(Phone) {}  
+    Person(string name, Date dob, int ID, long long int Phone): name(name), DOB(dob), ID(ID), Phone(Phone) {}  
 
     string getName() const {
         return name;
